Support more coefficients in CInet::Relation
We now can represent oriented and partially defined structures.

Right now, this comes at the expense of not having join and meet
implemented anymore. I'm not sure if that is even consistently
possible in full generality. I will have to think about it.

It is also not clear if making CInet::Relation the one class which
deals with all types of coefficients will not create problems later
when self-clocking code sees too late that it has to operate on
oriented or partially defined data.

There is also a risk that every implementation downstream of a
relation will have too many cases to support, like join and meet.