Add optional memory test
I'm trying to reduce memory usage and suspect there is even a leak when
using CInet::Seq's. At the end of the memory test, we have

  Orbit counting ->representative
    SCALAR                            8458204
    REF                                343242
    CInet::Hash::FaceKey               336631
    REF-CInet::Hash::FaceKey           336631
    ARRAY                                8275

  Files=7, Tests=58, 78 wallclock secs ( 0.02 usr  0.02 sys + 77.26 cusr 1.09 csys = 78.39 CPU)

The number of scalars is huge because each CInet::Relation is an arrayref
of small scalars. This should be improved first. Then, I'm worried about
the usage never decreasing, but maybe that is just how Perl works?
Between the last two runs, the number of scalars does not increase much,
indicating that allocated memory is never freed but re-used, so not really
leaking in the sense of a circular reference bug in Seq.