Rename some unused methods, add ->cival
Rename ->independent and ->dependent to ->independences and ->dependences
to highlight that lists are returned, not booleans.

The new method ->cival is an lvalue which can be used to set the value
of a CI statement. CInet::Graphical used the previous array access to
manipulate a CI structure. This method is a substitute for the recently
removed array interface.