Relation->new: use '1' for non-specified statements
The constructor for CInet::Relation receiving a list of CI statements
would create a relation which is undefined (having '*' coefficients)
for all statements that are not mentioned. But this seems counter-
intuitive since often relations are meant as sets. So now the relation
is dependent (coefficient '1') wherever not explicitly specified.

The '*' semantics can still be achieved easily by making an empty CIR
and setting the given statements using `cival`. In the future I might
add methods `weak` and `strong` (but with better names) which switch
all dependencies to undefinedness and vice versa.

Version bump for this incompatible change.